Don't override unexpected err with pub status
Previously, this line would override a close() due to Unknown Errors
from the TCPROSTransport connect fn (tcpros_base.py::connect():L567)
by setting the conn.done to the state of the publisher endpoint.
This would cause errors with the subscriber (self.protocol == None)
to get stuck in an infinite loop and cause a memory leak.

Not sure setting the done to True here is the right approach anyways
and should prob make the self.close() fn here be able to be called
multiple times, and clean-up from race conditions if need be, and
use a different "reset_conn" fn or different condition/flag for the
check_if_still_publisher case.